+++ src/PSRule/PSRule.psm1
@@ -14,6 +14,9 @@ Set-StrictMode -Version latest;
 #
 
 Import-LocalizedData -BindingVariable LocalizedHelp -FileName 'PSRule.Resources.psd1' -ErrorAction SilentlyContinue;
+if ($Null -eq (Get-Variable -Name LocalizedHelp -ErrorAction SilentlyContinue)) {
+    Import-LocalizedData -BindingVariable LocalizedHelp -FileName 'PSRule.Resources.psd1' -UICulture 'en-US' -ErrorAction SilentlyContinue;
+}
 
 #
 # Public functions
@@ -973,6 +976,10 @@ function New-PSRuleOption {
         [ValidateSet('Detail', 'Summary')]
         [PSRule.Configuration.ResultFormat]$OutputAs = 'Detail',
 
+        # Sets the Output.Culture option
+        [Parameter(Mandatory = $False)]
+        [String[]]$OutputCulture,
+
         # Sets the Output.Encoding option
         [Parameter(Mandatory = $False)]
         [ValidateSet('Default', 'UTF8', 'UTF7', 'Unicode', 'UTF32', 'ASCII')]
@@ -1154,6 +1161,10 @@ function Set-PSRuleOption {
         [Parameter(Mandatory = $False)]
         [ValidateSet('Detail', 'Summary')]
         [PSRule.Configuration.ResultFormat]$OutputAs = 'Detail',
+
+        # Sets the Output.Culture option
+        [Parameter(Mandatory = $False)]
+        [String[]]$OutputCulture,
 
         # Sets the Output.Encoding option
         [Parameter(Mandatory = $False)]
@@ -1739,6 +1750,10 @@ function SetOptions {
         [ValidateSet('Detail', 'Summary')]
         [PSRule.Configuration.ResultFormat]$OutputAs = 'Detail',
 
+        # Sets the Output.Culture option
+        [Parameter(Mandatory = $False)]
+        [String[]]$OutputCulture,
+
         # Sets the Output.Encoding option
         [Parameter(Mandatory = $False)]
         [ValidateSet('Default', 'UTF8', 'UTF7', 'Unicode', 'UTF32', 'ASCII')]
@@ -1840,6 +1855,11 @@ function SetOptions {
         # Sets option Output.As
         if ($PSBoundParameters.ContainsKey('OutputAs')) {
             $Option.Output.As = $OutputAs;
+        }
+
+        # Sets option Output.Culture
+        if ($PSBoundParameters.ContainsKey('OutputCulture')) {
+            $Option.Output.Culture = $OutputCulture;
         }
 
         # Sets option Output.Encoding